test: 테스트 코드 컨벤션 수정 및 파일 분리, 폴더링
